+++ public/pages/js/article.js
@@ -11,6 +11,10 @@ deleteBtns.forEach( deleteBtn=>{
 });
 
 async function deleteComment(articleId, commentId) {
+    result = confirm("Are you sure? If you delete this comment, you cannot recover it");
+    if(!result) {
+       return;
+    }
     const response = await fetch(`/articles/${articleId}/delete-comment/${commentId}`,{
         method : "DELETE"
     })
@@ -18,13 +22,35 @@ async function deleteComment(articleId, commentId) {
     const data = await response.json();
     console.log(data);
     if(data.success){
-        result = confirm("Are you sure? If you delete this comment, you cannot recover it");
-        if(result) {
-            console.log(commentId)
-            document.getElementById(`comments${commentId}`).remove();
-        }
+        console.log(commentId)
+        document.getElementById(`comments${commentId}`).remove();
     } else {
         alert(data.message);
     }
     console.log("mama", data.success, "papa", data.message);
+}
+
+const changeBtns = document.querySelectorAll(`.changeBtn`);
+
+changeBtns.forEach(changeBtn => {
+   changeBtn.addEventListener(`click`, () => {
+       const commentId = changeBtn.dataset.commentId;
+       console.log(commentId);
+   });
+});
+
+async function changeComment(commentId, newText, authorName) {
+    const response = await fetch('',{
+        method : "PUT",
+        body : JSON.stringify({commentId: commentId, text: newText, author: authorName})
+    });
+
+    const data = response.json();
+    console.log(data);
+
+    if(data.success){
+
+    } else {
+
+    }
 }
